+++ src/Doctrine/Util/Limiter.php
@@ -15,6 +15,41 @@ class Limiter
     private $total;
 
     /**
+     * @param int $tableCount1
+     * @param int $tableCount2
+     * @param int $limit
+     * @param int $offset
+     *
+     * @return array<string,int>
+     *
+     * @api
+     */
+    public static function getTwoTablePagination($tableCount1, $tableCount2, $limit, $offset = 0)
+    {
+        $offset1 = $offset > $tableCount1 ? $tableCount1 : $offset;
+        $limit1 = min($limit, $tableCount1 - $offset1);
+        if ($offset1 + $limit1 < $tableCount1) {
+            $offset2 = 0;
+            $limit2 = 0;
+        } elseif ($offset >= $tableCount1 + $tableCount2) {
+            $offset2 = $tableCount2;
+            $limit2 = 0;
+        } elseif ($offset < $tableCount1) {
+            $offset2 = 0;
+            $limit2 = $offset1 + $limit - $tableCount1;
+        } else {
+            $offset2 = $offset - $tableCount1;
+            $limit2 = $limit;
+        }
+        return [
+            'limit1'  => $limit1,
+            'offset1' => $offset1,
+            'limit2'  => $limit2,
+            'offset2' => $offset2,
+        ];
+    }
+
+    /**
      * @param int $limit how many records to get (e.g. LIMIT part of SQL queries)
      * @param int $offset how many records to skip (e.g. OFFSET part of SQL queries)
      * @param int|null $total how many total records were found (e.g. SELECT COUNT(*) FROM)