+++ src/Monolog/Util/Dumper.php
@@ -16,7 +16,7 @@ class Dumper
      */
     public static function dumpToString($value)
     {
-        $string = (string) print_r($value, true);
+        $string = (string)print_r($value, true);
         while (preg_match('/\n|\s{2}/', $string)) {
             $string = preg_replace('/\n\s*/', '', $string);
             $string = preg_replace('/\s{2,}/', ' ', $string);